[REFACTOR] section 폴더 삭제
