+++ src/screen/Main.js
@@ -2,14 +2,14 @@ import { Box, Container, Unstable_Grid2 as Grid } from "@mui/material";
 import SideNav from "../component/common/side-nav/SideNav";
 import { subDays, subHours } from "date-fns";
 import { MainContainer } from "../component/common/MainComponent";
-import { OverviewBudget } from "../component/main/section/OverviewBudget";
-import { OverviewTotalCustomers } from "../component/main/section/OverviewTotalCustomers";
-import { OverviewTotalProfit } from "../component/main/section/OverviewTotalProfit";
-import { OverviewTasksProgress } from "../component/main/section/OverviewTasksProgress";
-import { OverviewSales } from "../component/main/section/OverviewSales";
-import { OverviewTraffic } from "../component/main/section/OverviewTraffic";
-import { OverviewLatestProducts } from "../component/main/section/OverviewLatestProducts";
-import { OverviewLatestOrders } from "../component/main/section/OverviewLatestOrders";
+import { OverviewBudget } from "../component/main/OverviewBudget";
+import { OverviewTotalCustomers } from "../component/main/OverviewTotalCustomers";
+import { OverviewTotalProfit } from "../component/main/OverviewTotalProfit";
+import { OverviewTasksProgress } from "../component/main/OverviewTasksProgress";
+import { OverviewSales } from "../component/main/OverviewSales";
+import { OverviewTraffic } from "../component/main/OverviewTraffic";
+import { OverviewLatestProducts } from "../component/main/OverviewLatestProducts";
+import { OverviewLatestOrders } from "../component/main/OverviewLatestOrders";
 import { ACCESS_TOKEN, REFRESH_TOKEN } from "../constants/token";
 import { useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
